Fix HistoryService token user id extraction

+++ src/main/java/ticketsystem/DomainLayer/IRepository/IHistoryRepository.java
@@ -7,7 +7,7 @@ import ticketsystem.DomainLayer.history.Purchase;
 public interface IHistoryRepository {
     void addPurchase(Purchase purchase); 
     Purchase findPurchaseById(int purchaseId);
-    List<Purchase> getPurchasesByMemberId(int memberId);
+    List<Purchase> getPurchasesByMemberId(long memberId);
     List<Purchase> getPurchasesByCompanyId(int companyId);
     List<Purchase> getAllPurchases();
     int generateNextId();
